+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 3.9.6
 
+- Improvement: Added `Partial` to `CSharpClassMethod`.
 - Improvement: Added `AddNestedRecord` to `CSharpClass`.
 - Improvement: Clears InteractionStrategy cache between "Hot" Software Factory restarts.
 - Improvement: Added `Blank line between Members` setting to allow for "always" adding a blank line between members.